radar ukf

+++ src/main.cpp
@@ -76,7 +76,7 @@ int main() {
 						iss >> py;
 						meas_package.raw_measurements_ << px, py;
 						iss >> timestamp;
-						meas_package.timestamp_ = timestamp;
+						meas_package.timestamp_ = (long)timestamp;
 					} else if (sensor_type.compare("R") == 0) {
 
 						meas_package.sensor_type_ = MeasurementPackage::RADAR;
@@ -89,7 +89,7 @@ int main() {
 						iss >> ro_dot;
 						meas_package.raw_measurements_ << ro, theta, ro_dot;
 						iss >> timestamp;
-						meas_package.timestamp_ = timestamp;
+						meas_package.timestamp_ = (long)timestamp;
 					}
 					float x_gt;
 					float y_gt;
